Add one click deploy to Heroku and instructions
