+++ app.json
@@ -1,0 +1,11 @@
+{
+  "name": "SendGrid Inbound Parse",
+  "description": "Consume and parse POSTs from SendGrid's Inbound Parse Webhook",
+  "keywords": [
+    "sendgrid",
+    "inbound parse"
+  ],
+  "website": "http://www.sendgrid.com",
+  "repository": "https://github.com/sendgrid/sendgrid-go/tree/heroku-deployment",
+  "logo": "https://uiux.s3.amazonaws.com/2016-logos/email-logo%402x.png"
+}
