+++ conf.json
@@ -1,4 +1,4 @@
 {
     "endpoint": "/inbound",
-    "port": ":8000"
+    "port": "8000"
 }
