+++ helpers/inbound/README.md
@@ -20,15 +20,14 @@ cd sendgrid-go
 Run the Inbound Parse listener in your terminal:
 
 ```bash
-cd helpers/inbound/
-go run inbound.go
+go build
+./sendgrid-go
 ```
 
 In another terminal, run the test data sender:
 
 ```bash
-cd [path to sendgrid-go/helpers/inbound]
-go run inbound.go ./sample_data/default_data.txt http://127.0.0.1:8000/inbound
+./sendgrid-go ./helpers/inbound/sample_data/default_data.txt http://localhost:8000/inbound
 ```
 
 More sample data can be found [here](https://github.com/sendgrid/sendgrid-go/tree/master/helpers/inbound/sample_data).
@@ -44,7 +43,8 @@ Run the Inbound Parse listener in your terminal:
 
 ```bash
 git clone https://github.com/sendgrid/sendgrid-go.git
-go run inbound.go
+go build
+./sendgrid-go
 ```
 
 In another terminal, use [ngrok](https://ngrok.com/) to allow external access to your machine:
@@ -62,7 +62,40 @@ Next, send an email to [anything]@inbound.yourdomain.com, then look at the termi
 <a name="heroku"></a>
 # Deploy to Heroku
 
-If you would like to help implement this, please see [this ticket](https://github.com/sendgrid/sendgrid-go/issues/85).
+Get a [Heroku](https://www.heroku.com) account.
+
+[![Deploy](https://www.herokucdn.com/deploy/button.svg)](https://heroku.com/deploy?template=https://github.com/sendgrid/sendgrid-go/tree/heroku-deployment)
+
+[Setup your MX records.](https://sendgrid.com/docs/Classroom/Basics/Inbound_Parse_Webhook/setting_up_the_inbound_parse_webhook.html#-Setup) Depending on your domain name host, you may need to wait up to 48 hours for the settings to propagate.
+
+Update your SendGrid Incoming Parse settings: [Settings Page](https://app.sendgrid.com/settings/parse) | [Docs](https://sendgrid.com/docs/Classroom/Basics/Inbound_Parse_Webhook/setting_up_the_inbound_parse_webhook.html#-Pointing-to-a-Hostname-and-URL)
+
+- For the HOSTNAME field, use the domain that you changed the MX records (e.g. inbound.yourdomain.com)
+- For the URL field, use the URL generated by Heroku + /inbound, e.g https://[name-of-your-app].herokuapp.com/inbound
+
+Next, send an email to [anything]@inbound.yourdomain.com, then look at your Heroku logs: https://dashboard.heroku.com/apps/[name-of-your-app]/logs
+
+While you are waiting for your MX records to propagate, you can test by using the test data sender:
+
+```bash
+git clone https://github.com/sendgrid/sendgrid-go.git
+cd sendgrid-go
+go build
+./sendgrid-go ./helpers/inbound/sample_data/default_data.txt https://[name-of-your-app].herokuapp.com/inbound
+```
+
+To make changes: clone, modify and push the changes. [Heroku CLI](https://devcenter.heroku.com/articles/heroku-command-line) is required.
+
+For example:
+```bash
+git clone [repo url]
+git checkout heroku-deployment
+heroku git:remote -a [name-of-your-app]
+---make changes---
+git add -A
+git commit -m "updated configuration"
+git push heroku heroku-deployment:master
+```
 
 <a name="code_walkthrough"></a>
 # Code Walkthrough
