+++ helpers/inbound/inbound_test.go
@@ -1,4 +1,4 @@
-package main
+package inbound
 
 import (
 	"io/ioutil"
